detach height
