+++ cmd/vaporcli/commands/vaporcli.go
@@ -143,7 +143,6 @@ func AddCommands() {
 	VaporcliCmd.AddCommand(submitTransactionCmd)
 	VaporcliCmd.AddCommand(estimateTransactionGasCmd)
 
-	VaporcliCmd.AddCommand(rollbackCmd)
 	VaporcliCmd.AddCommand(getBlockCountCmd)
 	VaporcliCmd.AddCommand(getBlockHashCmd)
 	VaporcliCmd.AddCommand(getBlockCmd)
